requirements.txt
exchangelib>=5.0.0
msal>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
